+++ app/src/main/contrib/components/Responsibilities/ModelRunParameters/ModelRunParametersUploadForm.tsx
@@ -1,62 +1,62 @@
 import * as React from "react";
-import fetcher from "../../../../shared/sources/Fetcher";
-import {UploadState} from "../BurdenEstimates/UploadForm";
+import {UploadForm} from "../../../../shared/components/UploadForm";
 
-const formStyles = require("../../../../shared/styles/forms.css");
-const buttonStyles = require("../../../../shared/styles/buttons.css");
-const styles = require("../Responsibilities.css");
-
-interface UploadFormProps {
+interface ModelRunparametersUploadFormProps {
     groupId: string;
     touchstoneId: string;
     diseases: string[];
     token: string;
 }
 
-export class ModelRunParametersUploadForm extends React.Component<UploadFormProps, UploadState> {
+interface ModelRunParametersUploadFormState {
+
+    diseaseSelected: boolean;
+    description: boolean;
+}
+
+export class ModelRunParametersUploadForm extends React.Component<ModelRunparametersUploadFormProps, ModelRunParametersUploadFormState> {
 
     constructor() {
         super();
         this.state = {
-            fileSelected: false,
-            fileName: ""
+            diseaseSelected: false,
+            description: false
         };
     }
 
-    handleChange(e: any) {
+    handleDiseaseChange(e: any) {
         this.setState({
-            fileSelected: true,
-            fileName: e.target.value.replace("C:\\fakepath\\", "")
+            diseaseSelected: e.target.value.length > 0
+        });
+    }
+
+    handleDescriptionChange(e: any) {
+        this.setState({
+            description: e.target.value.length > 1
         });
     }
 
     render() {
 
         const uploadText = "Choose a new model run parameter set";
-        const props = this.props;
-        const url = fetcher.fetcher.buildOneTimeLink(props.token);
-        const enabled = props.token != null;
-        const diseases = this.props.diseases.map((disease) => <option key={disease} value={disease}>{disease}</option>);
+        const diseases = this.props.diseases.map((disease) =>
+            <option key={disease} value={disease}>{disease}</option>);
 
         return <div>
-            <form action={url} className={formStyles.form}
-                  method="POST" encType="multipart/form-data">
+            <UploadForm token={this.props.token} uploadText={uploadText}
+                        enableSubmit={this.state.description && this.state.diseaseSelected}>
                 <div className="form-group">
-                    <label className={formStyles.customFileUpload}>
-                        <input name="file" type="file" onChange={this.handleChange.bind(this)}/>
-                        <div className={`${buttonStyles.button} ${styles.button} mt-2 mb-2`}>
-                            {uploadText}
-                        </div>
-                        <div className="mr-5">{this.state.fileSelected ? "File selected: " + this.state.fileName : ""}</div>
-                    </label>
+                    <select name="disease" className="form-control" onChange={this.handleDiseaseChange.bind(this)}>
+                        <option value="">-- Select a disease --</option>
+                        {diseases}
+                    </select>
                 </div>
                 <div className="form-group">
-                    <select name="disease">{diseases}</select>
+                    <input type="text" className="form-control" name="description"
+                           onChange={this.handleDescriptionChange.bind(this)}/>
                 </div>
-                <button type="submit"
-                        disabled={!this.state.fileSelected || !enabled}>Upload
-                </button>
-            </form>
+            </UploadForm>
+
         </div>;
     }
 }
